+++ odoo/js/functions/common.js
@@ -568,6 +568,7 @@ odoo.define("terminal.functions.Common", function (require) {
                 .then((result) => {
                     this.screen.updateInputInfo(login);
                     this.screen.print(`Successfully logged as '${login}'`);
+                    this.executeCommand("reload", false, true);
                     return result;
                 });
         },
@@ -576,6 +577,7 @@ odoo.define("terminal.functions.Common", function (require) {
             return session.session_logout().then((result) => {
                 this.screen.updateInputInfo("Public User");
                 this.screen.print("Logged out");
+                this.executeCommand("reload", false, true);
                 return result;
             });
         },
@@ -1148,23 +1150,14 @@ odoo.define("terminal.functions.Common", function (require) {
 
         //
         _onBusNotification: function (notifications) {
-            const NotifDatas = Object.values(notifications.data);
-            const l = NotifDatas.length;
+            const notif_datas = Object.values(notifications.data);
+            const l = notif_datas.length;
             for (let x = 0; x < l; ++x) {
-                const notif = NotifDatas[x];
+                const notif = notif_datas[x];
                 this.screen.print(
-                    "<strong>[<i class='fa fa-envelope-o'></i>] New Longpolling Notification:</stron>"
+                    `<strong>[<i class='fa fa-envelope-o'></i>][${moment().format()}] New Longpolling Notification:</stron>`
                 );
-                if (notif[0] !== "not") {
-                    this.screen.print(
-                        [
-                            `From: ${notif[0][0]}`,
-                            `Channel: ${notif[0][1]}`,
-                            `To: ${notif[0][2]}`,
-                        ],
-                        true
-                    );
-                }
+                this.screen.print([`Channel ID: ${JSON.stringify(notif[0])}`]);
                 this.screen.print(notif[1], false);
             }
         },